# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/annonces": {
      "post": {
        "tags": [
          "Annonce"
        ],
        "summary": "Create",
        "operationId": "create_api_annonces_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnonceCreateSchemas"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnonceCreateSchemasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Annonce"
        ],
        "summary": "Index",
        "operationId": "index_api_annonces_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search_date_depart",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Date Depart"
            }
          },
          {
            "name": "search_date_arrivee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Date Arrivee"
            }
          },
          {
            "name": "search_origin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Origin"
            }
          },
          {
            "name": "search_destination",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Destination"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnonceGetSchemasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/annonce/{id}": {
      "post": {
        "tags": [
          "Annonce"
        ],
        "summary": "Delete",
        "operationId": "delete_api_annonce__id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnonceDeleteSchemaResponce"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Annonce"
        ],
        "summary": "Show",
        "operationId": "show_api_annonce__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AnnonceShowSchemaResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Show Api Annonce  Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Annonce"
        ],
        "summary": "Update",
        "operationId": "update_api_annonce__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnonceUpdateSchemas"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AnnonceUpdateSchemaResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Update Api Annonce  Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/annonces_by_gp": {
      "get": {
        "tags": [
          "Annonce"
        ],
        "summary": "Indexbygp",
        "operationId": "indexByGp_api_annonces_by_gp_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gp_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 8,
              "title": "Gp Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnonceGetSchemasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reservations": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "summary": "Index",
        "operationId": "index_api_reservations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Reservation"
        ],
        "summary": "Create",
        "operationId": "create_api_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnnonceCreateSchemas": {
        "properties": {
          "kilos_disponibles": {
            "type": "integer",
            "title": "Kilos Disponibles"
          },
          "date_depart": {
            "type": "string",
            "format": "date-time",
            "title": "Date Depart"
          },
          "date_arrivee": {
            "type": "string",
            "format": "date-time",
            "title": "Date Arrivee"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "prix_du_kilo": {
            "type": "number",
            "title": "Prix Du Kilo"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          }
        },
        "type": "object",
        "required": [
          "kilos_disponibles",
          "date_depart",
          "date_arrivee",
          "description",
          "prix_du_kilo",
          "origin",
          "destination"
        ],
        "title": "AnnonceCreateSchemas"
      },
      "AnnonceCreateSchemasData": {
        "properties": {
          "gp_id": {
            "type": "integer",
            "title": "Gp Id"
          },
          "kilos_disponibles": {
            "type": "integer",
            "title": "Kilos Disponibles"
          },
          "date_depart": {
            "type": "string",
            "format": "date-time",
            "title": "Date Depart"
          },
          "date_arrivee": {
            "type": "string",
            "format": "date-time",
            "title": "Date Arrivee"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "prix_du_kilo": {
            "type": "number",
            "title": "Prix Du Kilo"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          }
        },
        "type": "object",
        "required": [
          "gp_id",
          "kilos_disponibles",
          "date_depart",
          "date_arrivee",
          "description",
          "prix_du_kilo",
          "origin",
          "destination"
        ],
        "title": "AnnonceCreateSchemasData"
      },
      "AnnonceCreateSchemasResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "annonce": {
            "$ref": "#/components/schemas/AnnonceCreateSchemasData"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "annonce"
        ],
        "title": "AnnonceCreateSchemasResponse"
      },
      "AnnonceDeleteSchemaResponce": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "AnnonceDeleteSchemaResponce"
      },
      "AnnonceGetSchemasDatas": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kilos_disponibles": {
            "type": "integer",
            "title": "Kilos Disponibles"
          },
          "date_depart": {
            "type": "string",
            "format": "date-time",
            "title": "Date Depart"
          },
          "date_arrivee": {
            "type": "string",
            "format": "date-time",
            "title": "Date Arrivee"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "prix_du_kilo": {
            "type": "number",
            "title": "Prix Du Kilo"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "gp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kilos_disponibles",
          "date_depart",
          "date_arrivee",
          "description",
          "prix_du_kilo",
          "origin",
          "destination",
          "gp"
        ],
        "title": "AnnonceGetSchemasDatas"
      },
      "AnnonceGetSchemasResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "annonces": {
            "items": {
              "$ref": "#/components/schemas/AnnonceGetSchemasDatas"
            },
            "type": "array",
            "title": "Annonces"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "annonces"
        ],
        "title": "AnnonceGetSchemasResponse"
      },
      "AnnonceShowSchemaData": {
        "properties": {
          "gp": {
            "$ref": "#/components/schemas/UserRegisterResponseData"
          },
          "kilos_disponibles": {
            "type": "integer",
            "title": "Kilos Disponibles"
          },
          "date_depart": {
            "type": "string",
            "format": "date-time",
            "title": "Date Depart"
          },
          "date_arrivee": {
            "type": "string",
            "format": "date-time",
            "title": "Date Arrivee"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "prix_du_kilo": {
            "type": "number",
            "title": "Prix Du Kilo"
          },
          "prix_par_piece": {
            "type": "number",
            "title": "Prix Par Piece"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          }
        },
        "type": "object",
        "required": [
          "gp",
          "kilos_disponibles",
          "date_depart",
          "date_arrivee",
          "description",
          "prix_du_kilo",
          "prix_par_piece",
          "origin",
          "destination"
        ],
        "title": "AnnonceShowSchemaData"
      },
      "AnnonceShowSchemaResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "annonce": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnnonceShowSchemaData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "annonce"
        ],
        "title": "AnnonceShowSchemaResponse"
      },
      "AnnonceUpdateSchemaResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "annonce": {
            "$ref": "#/components/schemas/AnnonceUpdateSchemasData"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "annonce"
        ],
        "title": "AnnonceUpdateSchemaResponse"
      },
      "AnnonceUpdateSchemas": {
        "properties": {
          "kilos_disponibles": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kilos Disponibles"
          },
          "date_depart": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Depart"
          },
          "date_arrivee": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Arrivee"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "prix_du_kilo": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prix Du Kilo"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          }
        },
        "type": "object",
        "title": "AnnonceUpdateSchemas"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReservationCreate": {
        "properties": {
          "annonce_id": {
            "type": "integer",
            "title": "Annonce Id"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReservationItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items",
            "default": []
          },
          "special_items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReservationSpecialItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "annonce_id"
        ],
        "title": "ReservationCreate"
      },
      "ReservationItemCreate": {
        "properties": {
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "item_name",
          "weight"
        ],
        "title": "ReservationItemCreate"
      },
...
      "ReservationSchemaResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "reservation": {
            "$ref": "#/components/schemas/ReservationSchemaData"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "reservation"
        ],
        "title": "ReservationSchemaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}